move AnvilCostFix to new improvements directory
